# internal/web/static/openapi.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API the Tunnel Manager screens are built on. Every path\nbelow sits under /api and needs a session, except POST\n/api/login and GET /api/setup.\n\nLogging in sets two cookies. Over HTTPS they are named\n__Host-tm_session and __Host-tm_csrf, and over plain HTTP they\nare tm_session and tm_csrf. A browser sends them back on its\nown, so \"Try it out\" on this page works as soon as you are\nlogged in to the screens in the same browser.\n\nEvery POST, PUT and DELETE also needs the token of the session\nin an X-CSRF-Token header. POST /api/login answers with it in\ndata.csrf_token; paste it into Authorize above and this page\nsends it. A GET needs no token.\n\nEvery answer has the same shape: {\"success\":true,\"data\":...}\nor {\"success\":false,\"error\":\"...\"}. An answer that says no\ncarries error_code, which names the refusal and is what a\nscript should match on, and error_args, the values its English\nsentence was written with.",
        "title": "Tunnel Manager API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "3.8.3"
    },
    "basePath": "/api",
    "paths": {
...
        "/uninstall": {
            "post": {
                "security": [
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Takes the account password. The tunnels come down, the files this installation is made of are removed, and the answer is written before the process goes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "uninstall"
                ],
                "summary": "Remove the installation and end the process",
                "parameters": [
                    {
                        "description": "The account password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.uninstallRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.uninstallResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "The password does not open this account. Nothing was stopped and nothing was removed",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    }
                }
            }
        },
...
        "/update/install": {
            "post": {
                "security": [
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Takes the account password. It answers that the install started and never that it finished: what the install ends with is a restart of the service answering the request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Start the install of the newest release",
                "parameters": [
                    {
                        "description": "The account password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateInstallRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.updateView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "The password does not open this account",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    },
                    "409": {
                        "description": "There is nothing newer, or this build cannot install one",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.errorArgs": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "api.errorBody": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "error_args": {
                    "$ref": "#/definitions/api.errorArgs"
                },
                "error_code": {
                    "$ref": "#/definitions/api.errorCode"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.removedFile": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error is filled in only for a file that could not be removed.",
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "what": {
                    "type": "string"
                },
                "what_code": {
                    "description": "WhatCode names What, so that a screen says it in the language it is\ndrawn in instead of showing the English above.",
                    "type": "string"
                }
            }
        },
...
        "api.uninstallRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "api.uninstallResult": {
            "type": "object",
            "properties": {
                "exit_in_sec": {
                    "description": "ExitInSec is how long the process stays up after this answer. The screen\nsays it, so the number comes from the side that waits it out.",
                    "type": "integer"
                },
                "failed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.removedFile"
                    }
                },
                "removed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.removedFile"
                    }
                }
            }
        },
        "api.updateInstallRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "api.updateView": {
            "type": "object",
            "properties": {
                "checked_at": {
                    "description": "Checked is when the last look happened, and the three fields under it are\nwhat it found. Checked is the zero time where nothing has looked yet,\nwhich the screen draws as never rather than as 1970.",
                    "type": "string"
                },
                "comparable": {
                    "type": "boolean"
                },
                "installable": {
                    "description": "Installable says whether the press that installs is offered. See the\nfield of the same name on the handler.",
                    "type": "boolean"
                },
                "newer": {
                    "type": "boolean"
                },
                "problem": {
                    "description": "Problem is why the last look found nothing, empty where it worked. A\ncheck that failed is not a version that is up to date, and this is what\nkeeps the two apart on the screen.",
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "version": {
                    "description": "Version is what this process is running, without the leading v, the way\nthe version constant carries it.",
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CSRFToken": {
            "description": "The token POST /api/login answers with, in\ndata.csrf_token. Required on every POST, PUT\nand DELETE. The session cookie goes with it and\nis sent by the browser on its own.",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}